+++ sc/social/like/plugins/telegram/browser.py
@@ -1,13 +1,12 @@
 # -*- coding:utf-8 -*-
-from Products.CMFPlone.utils import safe_unicode
-from plone.api.exc import InvalidParameterError
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.PythonScripts.standard import url_quote
+from plone.api.exc import InvalidParameterError
+from plone import api
 from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.utils import get_language
 from zope.component import getMultiAdapter
-from plone import api
 
 
 class PluginView(BrowserView):
@@ -21,23 +20,22 @@ class PluginView(BrowserView):
     def __init__(self, context, request):
         self.context = context
         self.request = request
-        # FIXME: the following could rise unexpected exceptions
-        #        move it to a new setup() method
-        #        see: http://docs.plone.org/develop/plone/views/browserviews.html#creating-a-view
+        self._setup()
+
+    def _setup(self):
         self.portal_state = getMultiAdapter(
             (self.context, self.request), name=u'plone_portal_state')
         self.portal = self.portal_state.portal()
         self.site_url = self.portal_state.portal_url()
         self.portal_title = self.portal_state.portal_title()
-        self.url = context.absolute_url()
-        self.language = get_language(context)
-        data = url_quote(u'{0} - {1}'.format(
-            safe_unicode(self.context.title), self.context.absolute_url()))
-        self.whatsappurl = u'whatsapp://send?text={0}'.format(data)
+        self.url = self.context.absolute_url()
+        self.language = get_language(self.context)
+        data = url_quote(self.context.absolute_url())
+        self.telegramurl = u'https://telegram.me/share/url?url={0}'.format(data)
 
     @property
     def klass(self):
-        klass = 'whatsapp'
+        klass = 'telegram'
         record = ISocialLikeSettings.__identifier__ + '.do_not_track'
         try:
             do_not_track = api.portal.get_registry_record(record)