+++ sc/social/like/tests/test_plugin_telegram.py
@@ -3,14 +3,14 @@ from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from sc.social.like.interfaces import ISocialLikeLayer
 from sc.social.like.plugins.interfaces import IPlugin
-from sc.social.like.plugins.whatsapp import browser
+from sc.social.like.plugins.telegram import browser
 from sc.social.like.testing import INTEGRATION_TESTING
 from zope.component import getUtilitiesFor
 from zope.interface import alsoProvides
 
 import unittest
 
-name = 'WhatsApp'
+name = 'Telegram'
 
 
 class PluginTest(unittest.TestCase):
@@ -28,7 +28,7 @@ class PluginTest(unittest.TestCase):
     def test_plugin_config(self):
         plugin = self.plugins[name]
         self.assertEqual(plugin.name, name)
-        self.assertEqual(plugin.id, 'whatsapp')
+        self.assertEqual(plugin.id, 'telegram')
 
     def test_plugin_config_view(self):
         plugin = self.plugins[name]
@@ -36,7 +36,7 @@ class PluginTest(unittest.TestCase):
 
     def test_plugin_view(self):
         plugin = self.plugins[name]
-        self.assertEqual(plugin.view(), '@@whatsapp-plugin')
+        self.assertEqual(plugin.view(), '@@telegram-plugin')
 
     def test_plugin_metadata(self):
         plugin = self.plugins[name]
@@ -77,7 +77,7 @@ class PluginViewsTest(unittest.TestCase):
         plugin_view = plugin.view()
         view = document.restrictedTraverse(plugin_view)
         html = view.plugin()
-        self.assertIn('whatsapp', html)
+        self.assertIn('telegram', html)
 
     def test_plugin_urlnoscript_encoding(self):
         plugin = self.plugin
@@ -88,4 +88,4 @@ class PluginViewsTest(unittest.TestCase):
         view = document.restrictedTraverse(plugin_view)
         html = view.plugin()
         self.assertIn(
-            'Not%C3%ADcia%20-%20http%3A//nohost/plone/my-document" class="whatsapp">Share', html)
+            'http%3A//nohost/plone/my-document" class="telegram">Share', html)