+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 2.4.2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Added Telegram plugin (closes `#52`_).
+  [rodfersou]
+
 - Use Plone's registry instead of the ``portal_properties`` tool to store package configuration (closes `#1`_).
   [hvelarde]
 
@@ -282,4 +285,5 @@ There's a frood who really knows where his towel is.
 .. _`#36`: https://github.com/collective/sc.social.like/issues/36
 .. _`#38`: https://github.com/collective/sc.social.like/issues/38
 .. _`#39`: https://github.com/collective/sc.social.like/issues/39
+.. _`#52`: https://github.com/collective/sc.social.like/issues/52
 .. _`#56`: https://github.com/collective/sc.social.like/issues/56